requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
chainlit==0.4.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==41.0.1
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.1
fastapi-socketio==0.0.10
fastjsonschema==2.17.1
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
google-search-results==2.4.2
googleapis-common-protos==1.59.1
grpcio==1.56.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.1
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.1
jupyter-events==0.6.3
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.202
langchainplus-sdk==0.0.10
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
setuptools==67.7.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
stack-data==0.6.2
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
uptrace==1.18.0
uri-template==1.2.0
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
websockets==11.0.3
wheel==0.40.0
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
